Utilize 'remove_root_dir_from_filename' when inserting or removing an entry from symbol database.
Usage of 'os.path.basename' instead was plain _WRONG_ because we are loosing
an information for files which are located in non-root directories of a project. I.e.
'os.path.basename' for '<path_to_project_dir>/filename1.cpp' will give us 'filename1.cpp'
back, and that is what we wanted, but when we run it on a '<path_to_project_dir>/lib/filename2.cpp'
we will get 'filename2.cpp' and that is not what we wante: we lost the 'lib/' part.