We don't want to insert symbol db entry with unique_id (USR) == ''.
This is a deficiency from libclang parsing step which sometimes may result
with an empty USR due to missing include path or similar information.

If we were to insert such an entry we would end up with many invalid entries
in the symbol database, and once we would make a query (i.e. find-all-references,
go-to-definition) over such a symbol in the source code, we would get a very
wrong result(s) back. It is better not to return anything (i.e. no-definition-found,
no-references) but to return a non-sense result back.